+++ skyperious/conf.py
@@ -23,7 +23,7 @@ import util
 
 """Program title, version number and version date."""
 Title = "Skyperious"
-Version = "3.6.dev0"
+Version = "3.6.dev1"
 VersionDate = "05.07.2020"
 
 if getattr(sys, "frozen", False):
@@ -52,7 +52,7 @@ OptionalFileDirectives = ["EmoticonsPlotWidth", "ExportChatTemplate",
     "MaxHistoryInitialMessages", "MaxRecentFiles", "MaxSearchHistory",
     "MaxSearchMessages", "MaxSearchTableRows", "PlotDaysColour",
     "PlotDaysUnitSize", "PlotHoursColour", "PlotHoursUnitSize",
-    "SearchContactsChunk", "SearchResultsChunk", "SharedImageAutoDownload",
+    "SearchContactsChunk", "SearchResultsChunk", 
     "StatisticsPlotWidth", "StatusFlashLength", "UpdateCheckInterval",
     "WordCloudLengthMin", "WordCloudCountMin", "WordCloudWordsMax",
     "WordCloudWordsAuthorMax"
@@ -319,11 +319,6 @@ StatisticsPlotWidth = 150
 
 """Width of the chat emoticons plots, in pixels."""
 EmoticonsPlotWidth = 200
-
-"""
-Ask for Skype password and download shared images from Skype web for HTML export.
-"""
-SharedImageAutoDownload = True
 
 """Duration of "flashed" status message on StatusBar, in milliseconds."""
 StatusFlashLength = 30000