+++ skyperious/main.py
@@ -19,7 +19,6 @@ import codecs
 import collections
 import datetime
 import errno
-import getpass
 import glob
 import locale
 import io
@@ -31,7 +30,6 @@ import sys
 import threading
 import time
 import traceback
-import warnings
 
 try:
     import wx
@@ -355,28 +353,12 @@ def run_search(filenames, query):
         worker and (worker.stop(), worker.join())
 
 
-def run_export(filenames, format, chatnames, authornames, ask_password):
+def run_export(filenames, format, chatnames, authornames):
     """Exports the specified databases in specified format."""
     dbs = [skypedata.SkypeDatabase(f) for f in filenames]
     is_xlsx_single = ("xlsx_single" == format)
 
     for db in dbs:
-        if (ask_password and db.id and conf.SharedImageAutoDownload
-        and format.lower().endswith("html")):
-            prompt = "Enter Skype password for '%s': " % db.id
-            while not skypedata.SharedImageDownload.has_login(db.id):
-                with warnings.catch_warnings():
-                    warnings.simplefilter("ignore") # possible GetPassWarning
-                    output(prompt, end="") # getpass output can raise errors
-                    pw = getpass.getpass("", io.BytesIO())
-                if not pw: continue # while
-                try:
-                    skypedata.SharedImageDownload.login(db.id, pw)
-                except Exception as e:
-                    log("Error signing in %s on Skype web.\n\n%s",
-                        db.id, util.format_exc(e))
-                    prompt = "%s\nEnter Skype password for '%s': " % (e, db.id)
-
         formatargs = collections.defaultdict(str)
         formatargs["skypename"] = os.path.basename(db.filename)
         formatargs.update(db.account or {})
@@ -569,8 +551,7 @@ def run(nogui=False):
     elif "merge" == arguments.command:
         run_merge(arguments.FILE, arguments.output)
     elif "export" == arguments.command:
-        run_export(arguments.FILE, arguments.type, arguments.chat,
-                   arguments.author, arguments.ask_password)
+        run_export(arguments.FILE, arguments.type, arguments.chat, arguments.author)
     elif "search" == arguments.command:
         run_search(arguments.FILE, arguments.QUERY)
     elif "gui" == arguments.command: