+++ skyperious/skypedata.py
@@ -18,7 +18,6 @@ import copy
 import cStringIO
 import csv
 import datetime
-import json
 import math
 import os
 import re
@@ -30,11 +29,8 @@ import textwrap
 import time
 import traceback
 import urllib
-import urllib2
 from xml.etree import cElementTree as ElementTree
 
-try: from bs4 import BeautifulSoup # For Skype shared photo login
-except ImportError: pass
 try: import wx # For avatar bitmaps in GUI program
 except ImportError: pass
 
@@ -1803,8 +1799,8 @@ class MessageParser(object):
                     url = dom.find("URIObject").get("uri")
             if url and self.stats:
                 self.stats["shared_images"][message["id"]] = dict(url=url,
-                    success=False, author_name=get_author_name(message),
-                    author=message["author"], datetime=message["datetime"])
+                    author_name=get_author_name(message), author=message["author"],
+                    datetime=message["datetime"])
             dom = self.sanitize(dom, ["a", "b", "i", "s", "ss", "quote", "span"])
 
         # Process Skype message quotation tags, assembling a simple
@@ -1893,13 +1889,6 @@ class MessageParser(object):
 
     def dom_to_html(self, dom, output, message):
         """Returns an HTML representation of the message body."""
-        shared_image = self.stats.get("shared_images", {}).get(message["id"])
-        if shared_image and conf.SharedImageAutoDownload and output.get("export"):
-            raw = SharedImageDownload.get_image(self.db.id, shared_image["url"])
-            if raw:
-                shared_image["success"] = True
-                ns = dict(shared_image, image=raw, message_id=message["id"])
-                return step.Template(templates.CHAT_MESSAGE_IMAGE).expand(ns)
         other_tags = ["blink", "font", "span", "table", "tr", "td", "br"]
         greytag, greyattr, greyval = "font", "color", conf.HistoryGreyColour
         if output.get("export"):
@@ -2295,116 +2284,6 @@ class MessageParser(object):
 
         return stats
 
-
-
-class SharedImageDownload(object):
-    """
-    Static class for maintaining Skype user login sessions to Skype website 
-    and downloading shared images.
-    """
-    LOGINS = {} # {username: True, }
-    OPENERS = {} # {username: urllib2.OpenerDirector, }
-    STARTURL = ("https://login.skype.com/login?application=asm&return_url="
-                "https%3A%2F%2Fapi.asm.skype.com%2Fv1%2Fskypetokenauth%3F"
-                "redirectUrl%3Dhttps%3A%2F%2Fapi.asm.skype.com%2Fs%2Fi%3F&"
-                "message=signin_continue")
-
-    @classmethod
-    def has_login(cls, username):
-        """Returns whether an active login session exists for username."""
-        return cls.LOGINS.get(username, False)
-
-
-    @classmethod
-    def get_image(cls, username, url):
-        """Returns image raw data, or None if not available."""
-        if not cls.has_login(username): return None
-        main.log("Retrieving shared image %s.", url)
-        content = None
-        try:
-            statusurl = (url.replace("skype.com/s/i?", "skype.com/v1/objects/")
-                         + "/views/imgo/status")
-            statuscontent, _ = cls.request(username, statusurl)
-            if statuscontent: # Actual image URL given in web service JSON
-                imagedata = json.loads(statuscontent)
-                if not imagedata or "view_location" not in imagedata:
-                    err = ("Response unrecognized from %s.\n\n%s." %
-                           (statusurl, statuscontent))
-                    main.log(err), support.report_error(err)
-                else:
-                    imageurl = imagedata["view_location"]
-                    content, _ = cls.request(username, imageurl)
-        except Exception:
-            err = ("Error retrieving %s from Skype web.\n\n%s" %
-                   (url, traceback.format_exc()))
-            main.log(err), support.report_error(err)
-        return content
-
-
-    @classmethod
-    def login(cls, username, password):
-        """Logs in to Skype online successfully or raises exception."""
-        main.log("Logging in to Skype web as '%s'.", username)
-        for dct in (cls.LOGINS, cls.OPENERS): dct.pop(username, None)
-
-        # 1. Retrieve Skype website login form variables.
-        content1, code1 = cls.request(username, cls.STARTURL)
-        if not content1:
-            raise Exception("Response %s from start page." % code1)
-        soup1 = BeautifulSoup(content1, "html.parser")
-        form1 = soup1.find("form", {"id": "loginForm"})
-        if not form1:
-            raise Exception("No login form detected on start page.")
-        loginurl = form1["action"]
-        logindata = dict((x["name"], x.get("value", ""))
-                          for x in form1.findAll("input"))
-        logindata.update(username=username, password=password)
-
-        # 2. Post login data, retrieve cookie form variables.
-        data2 = urllib.urlencode(logindata)
-        content2, code2 = cls.request(username, loginurl, data2)
-        if not content2:
-            raise Exception("Response %s from login page." % code2)
-        soup2 = BeautifulSoup(content2, "html.parser")
-        loginfailed = soup2.find("div", {"class": "messageBox message_error"})
-        if loginfailed:
-            elem = loginfailed.find("span")
-            raise Exception("Response '%s' from https://login.skype.com/." %
-                            (elem.text if elem else "LOGIN FAILED"))
-        form2 = soup2.find("form", {"id": "redirectForm"})
-        if not form2:
-            raise Exception("No redirect form detected on token page.")
-        tokenurl = form2["action"]
-        tokendata = dict((f["name"], f.get("value", ""))
-                          for f in form2.findAll("input"))
-
-        # 3. Post cookie token data to finalize login.
-        data3 = urllib.urlencode(tokendata)
-        content3, code3 = cls.request(username, tokenurl, data3)
-        if not content3:
-            raise Exception("Response %s from token page." % code3)
-        cls.LOGINS[username] = True
-
-
-    @classmethod
-    def request(cls, username, url, data=None):
-        """Returns (URL contents, HTTP code) or (None, None)."""
-        content, code = None, None
-        if username not in cls.OPENERS:
-            cookiehandler = urllib2.HTTPCookieProcessor(cookielib.CookieJar())
-            cls.OPENERS[username] = urllib2.build_opener(cookiehandler)
-        try:
-            resp = cls.OPENERS[username].open(url, data)
-            code, content = resp.code, resp.read()
-        except urllib2.HTTPError as e:
-            code = e.code
-            err = "Response %s from %s.\n\n%s" % (e.code, url, e.read())
-            main.log(err), support.report_error(err)
-        except urllib2.URLError:
-            err = ("Error retrieving %s from Skype web.\n\n%s" %
-                   (url, traceback.format_exc()))
-            main.log(err), support.report_error(err)
-        return content, code
 
 
 def is_skype_database(filename, path=None):